WIP: transition to native-base
