+++ package.json
@@ -87,6 +87,7 @@
     "i18next": "^21.8.4",
     "jest": "^26.6.3",
     "moti": "^0.18.0",
+    "native-base": "^3.4.7",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-hook-form": "^7.31.2",
