+++ src/constants/index.ts
@@ -3,3 +3,4 @@ export * from './colorSchemes'
 export * from './environments'
 export * from './regex'
 export * from './theme'
+export * from './NativeBaseConfig'
