+++ src/providers/index.ts
@@ -1,3 +1,4 @@
+export { NativeBaseProvider } from 'native-base'
 export { SafeAreaProvider } from 'react-native-safe-area-context'
 export * from './AuthProvider'
 export * from './ColorSchemeProvider'
