+++ src/screens/DetailsScreen.tsx
@@ -1,15 +1,11 @@
 import { BottomSheetModal, BottomSheetBackdrop } from '@gorhom/bottom-sheet'
+import { Box, Center, Text, Button } from 'native-base'
 import { useCallback, useMemo, useRef } from 'react'
-import { View } from 'react-native'
-
-import { Button, Text } from '~components'
-import { useTheme } from '~hooks'
 
 export const DetailsScreen = (props: DetailsScreenProps): JSX.Element => {
   const {
     route: { params },
   } = props
-  const { s } = useTheme()
 
   // ref
   const bottomSheetModalRef = useRef<BottomSheetModal>(null)
@@ -32,9 +28,9 @@ export const DetailsScreen = (props: DetailsScreenProps): JSX.Element => {
   )
 
   return (
-    <View style={[s.flex1, s.justifyCenter, s.itemsCenter]}>
+    <Center>
       <Text>This is details screen</Text>
-      <Button onPress={openModal} title="Open BottomSheetModal" style={[s.mB2]} />
+      <Button onPress={openModal}>Open BottomSheetModal</Button>
       <Text>Screen params {JSON.stringify(params)}</Text>
       <BottomSheetModal
         ref={bottomSheetModalRef}
@@ -43,10 +39,10 @@ export const DetailsScreen = (props: DetailsScreenProps): JSX.Element => {
         onChange={handleSheetChanges}
         backdropComponent={renderBackdrop}
       >
-        <View style={[s.p4]}>
+        <Box p={4}>
           <Text color="black">Awesome 🎉</Text>
-        </View>
+        </Box>
       </BottomSheetModal>
-    </View>
+    </Center>
   )
 }
