+++ src/services/index.ts
@@ -1,2 +1,3 @@
 export * from './MockedServerService'
 export * from './TokenService'
+export * from './ColorSchemeManager'
